src/main/java/application/Utils/DatabaseUtil.java
package application.Utils;

import java.sql.Connection;
import java.sql.DriverManager;
import java.sql.SQLException;

public class DatabaseUtil {

    private static final String URL = "jdbc:mysql://localhost:3306/lms";
    private static final String USERNAME = "root";
    private static final String PASSWORD = "Rwirasira@123";
    // Private constructor to prevent instantiation
    public DatabaseUtil() {}


    // Static method to get a database connection
    public static Connection getConnection() throws SQLException {
        try {
            Class.forName("com.mysql.cj.jdbc.Driver");
        } catch (ClassNotFoundException e) {
            throw new SQLException("MySQL JDBC driver not found", e);
        }

        return DriverManager.getConnection(URL, USERNAME, PASSWORD);
    }


    // Method to close a connection
    public static void close(Connection connection) {

        if (connection != null) {
            try {
                connection.close();
            } catch (SQLException e) {
                System.err.println("Error closing database connection: " + e.getMessage());
            }
        }
    }
}
